refactoring and new features
new OperationalGrid class to take SRGrid stuff and move it back in the inheritance chain since we have these new backends
add an is_backscatter property that will force layernames from depth to intensity, and prevent stuff like generating BAGs
added tests for this new tech
